requirements.txt
PySide6
psutil
requests
Pillow
extract-icon
pefilepyinstaller
